feat: créa° d'un input standalone pour les collectivités
